Reorganize deep_srsnv into pipeline-stage subpackages
Split the monolithic deep_srsnv module into domain-based folders:

- utils/vocab.py: Encoders, load_vocab_config, channel constants
- utils/alignment.py: CIGAR processing, gapped channel building
- utils/split_utils.py: fold assignment, split ID computation
- preprocessing/cram_to_tensors.py: CRAM tensorization CLI
- preprocessing/combine_splits.py: fold splitting CLI
- training/data_module.py: PyTorch Lightning data module
- training/lightning_module.py: Lightning training module
- training/cnn_model.py: CNN architecture
- training/swa_validation_tracker.py: SWA callback

data_prep.py is slimmed to backward-compat re-exports.
All entry points, inference/, and test imports updated.
