+++ src/srsnv/pyproject.toml
@@ -38,8 +38,8 @@ srsnv_inference = "ugbio_srsnv.srsnv_inference:main"
 srsnv_training = "ugbio_srsnv.srsnv_training:main"
 srsnv_dnn_bam_training = "ugbio_srsnv.srsnv_dnn_bam_training:main"
 srsnv_report = "ugbio_srsnv.srsnv_report:main"
-cram_to_tensors = "ugbio_srsnv.deep_srsnv.cram_to_tensors:main"
-combine_splits = "ugbio_srsnv.deep_srsnv.combine_splits:main"
+cram_to_tensors = "ugbio_srsnv.deep_srsnv.preprocessing.cram_to_tensors:main"
+combine_splits = "ugbio_srsnv.deep_srsnv.preprocessing.combine_splits:main"
 dnn_fold_inference_from_cache = "ugbio_srsnv.deep_srsnv.inference.dnn_fold_inference:main_fold_from_cache"
 dnn_merge_and_annotate = "ugbio_srsnv.deep_srsnv.inference.dnn_merge_annotate:main"
 prepare_dnn_report = "ugbio_srsnv.prepare_dnn_report:main"
