+++ src/srsnv/ugbio_srsnv/deep_srsnv/inference/export.py
@@ -10,7 +10,7 @@ import onnx
 import torch
 from ugbio_core.logger import logger
 
-from ugbio_srsnv.deep_srsnv.cnn_model import CNNReadClassifier
+from ugbio_srsnv.deep_srsnv.training.cnn_model import CNNReadClassifier
 
 
 def export_to_onnx(
