+++ src/srsnv/ugbio_srsnv/deep_srsnv/training/cnn_model.py
